// src/types/User.ts
export interface User {
    id?: number;
    gender: string;
    email: string;
    address: string;
    userType: string;
    status: string;
    username: string;
    first_name: string;
    last_name: string;
    date_of_birth: string;
    phone_number: string;
    created_at?: string;
    updated_at?: string;
    created_by?: string;
    updated_by?: string;
}